fix: move states to context
