+++ src/App.jsx
@@ -5,8 +5,10 @@ import CommentPage from "./Pages/CommentPage/CommentPage.jsx";
 import ErrorComponent from "./Pages/ErrorPage.jsx";
 import { AuthContextProvider } from "./context/authContext";
 import { CommentContextProvider } from "./context/commentContext.jsx";
+import { IssueContextProvider } from "./context/issueContext";
 import { ThemeProvider, BaseStyles } from "@primer/react";
 import PrivateRoute from "./utils/PrivateRoute.jsx";
+
 function App() {
   return (
     <ThemeProvider>
@@ -17,9 +19,11 @@ function App() {
             <Route
               path="/:owner/:repoName/issue"
               element={
-                <PrivateRoute>
-                  <IssuePage />
-                </PrivateRoute>
+                <IssueContextProvider>
+                  <PrivateRoute>
+                    <IssuePage />
+                  </PrivateRoute>
+                </IssueContextProvider>
               }
             />
             <Route
