+++ src/Pages/IssuePage/IssuePage.jsx
@@ -1,226 +1,31 @@
-import { useState, useEffect, useCallback } from "react";
-import api from "../../utils/api";
+import { useContext } from "react";
+import { IssueContext } from "../../context/issueContext";
 import { Center } from "../../style/Center.styled";
-import { useNavigate, useParams } from "react-router-dom";
 import IssueSearch from "./IssuePageSearch";
 import IssuePageHeader from "./IssuePageHeader";
 import IssuePageList from "./IssuePageList";
 import { IssueAllContainer } from "../../style/IssuePage.styled";
-
 import { Button } from "@primer/react";
 import { XCircleFillIcon } from "@primer/octicons-react";
 
 const IssuePage = () => {
-  const [apiResult, setApiResult] = useState([]);
-  const [allIssues, setAllIssues] = useState({ openCount: 0, closedCount: 0 });
-  const [authors, setAuthors] = useState([]);
-  const [labels, setLabels] = useState([]);
-  const [selectedAuthor, setSelectedAuthor] = useState("all");
-  const [selectedLabel, setSelectedLabel] = useState("all");
-  const [searchValue, setSearchValue] = useState("");
-  const [stateOpenOrClosed, setStateOpenOrClosed] = useState("open");
-  const { repoName, owner } = useParams();
-  const navigate = useNavigate();
-
-  const parseUrlParams = () => {
-    const searchParams = new URLSearchParams(window.location.search);
-    const query = searchParams.get("q") || "";
-
-    const params = query.split("+").reduce(
-      (acc, param) => {
-        if (param.startsWith("repo:")) {
-          acc.repo = param.substring(5);
-        } else if (param.startsWith("is:")) {
-          acc.state = param.substring(3);
-        } else if (param.startsWith("label:")) {
-          acc.labels = acc.labels || [];
-          acc.labels.push(param.substring(6));
-        } else if (param.startsWith("author:")) {
-          acc.author = param.substring(7);
-        }
-        return acc;
-      },
-      { labels: [] }
-    );
-
-    const searchResult = query
-      .split("+")
-      .filter(
-        (param) =>
-          !param.startsWith("repo:") &&
-          !param.startsWith("is:") &&
-          !param.startsWith("label:") &&
-          !param.startsWith("author:")
-      )
-      .join(" ");
-
-    return {
-      state: params.state || "open",
-      author: params.author || "all",
-      labels: params.labels || [],
-      searchResult: searchResult.trim(),
-    };
-  };
-
-  const fetchInitialData = useCallback(async () => {
-    try {
-      if (owner && repoName) {
-        const response = await api.fetchInitialData(owner, repoName);
-        setApiResult(response.issues);
-        setLabels(response.labels);
-        setAllIssues({
-          openCount: response.openCount,
-          closedCount: response.closedCount,
-        });
-        setAuthors(response.uniqueAuthors);
-      }
-    } catch (error) {
-      console.error("Failed to fetch initial data:", error);
-      const errorMessage = error.message || "Something went wrong";
-      navigate("/error", { state: { errorMessage } });
-    }
-  }, [navigate, owner, repoName]);
-
-  const fetchDataAndUpdateUrl = useCallback(() => {
-    if (!owner || !repoName || !stateOpenOrClosed) {
-      const errorMessage = "Repository owner, name, and state are required.";
-      navigate("/error", { state: { errorMessage } });
-      return;
-    }
-
-    const updateUrlParams = () => {
-      const url = new URL(window.location.href);
-      const searchParams = new URLSearchParams();
-
-      const state =
-        stateOpenOrClosed === "open"
-          ? "is:issue is:open"
-          : `is:issue is:${stateOpenOrClosed}`;
-      const author = selectedAuthor !== "all" ? `author:${selectedAuthor}` : "";
-      const label = selectedLabel !== "all" ? `${selectedLabel}` : "";
-      const searchResult = searchValue || "";
-
-      const queryString = [state, author, label, searchResult]
-        .filter(Boolean)
-        .join(" ");
-
-      if (queryString.trim() !== "is:issue is:open") {
-        searchParams.set("q", queryString);
-      } else {
-        searchParams.delete("q");
-      }
-      url.search = searchParams.toString();
-      window.history.pushState({}, "", url);
-    };
-
-    updateUrlParams();
-
-    const fetchFilteredIssues = async () => {
-      try {
-        const authorFilter = selectedAuthor || "all";
-        const labelFilter = selectedLabel || "";
-        const searchResult = searchValue || "";
-        const q = parseUrlParams();
-
-        const response = await api.fetchFilteredIssues(
-          q.searchResult,
-          owner,
-          repoName,
-          authorFilter,
-          labelFilter,
-          stateOpenOrClosed,
-          searchResult
-        );
-
-        setApiResult(response);
-      } catch (error) {
-        console.error("Failed to fetch filtered issues:", error);
-        const errorMessage = error.message || "Something went wrong";
-        navigate("/error", { state: { errorMessage } });
-      }
-    };
-
-    fetchFilteredIssues();
-  }, [
-    navigate,
+  const {
+    apiResult,
+    allIssues,
+    authors,
+    labels,
+    selectedAuthor,
+    selectedLabel,
+    searchValue,
+    stateOpenOrClosed,
     owner,
     repoName,
-    searchValue,
-    selectedAuthor,
-    selectedLabel,
-    stateOpenOrClosed,
-  ]);
-
-  useEffect(() => {
-    const { state, author, labels, searchResult } = parseUrlParams();
-    setSelectedAuthor(author);
-    setSelectedLabel(labels.join(" "));
-    setSearchValue(searchResult);
-    setStateOpenOrClosed(state);
-    fetchInitialData();
-  }, [fetchInitialData]);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      fetchDataAndUpdateUrl();
-    }, 500);
-
-    return () => clearTimeout(timer);
-  }, [
-    fetchDataAndUpdateUrl,
-    searchValue,
-    selectedAuthor,
-    selectedLabel,
-    stateOpenOrClosed,
-  ]);
-
-  const handleFilterChange = (type, value) => {
-    if (type === "label") {
-      setSelectedLabel(value);
-    } else if (type === "author") {
-      setSelectedAuthor(value);
-    }
-
-    fetchDataAndUpdateUrl();
-  };
-
-  const handleAuthorChange = (selectedAuthor) => {
-    handleFilterChange("author", selectedAuthor);
-  };
-
-  const handleLabelChange = (labels) => {
-    const formattedString =
-      labels.length > 0
-        ? labels.map((label) => `label:"${label}"`).join(" ")
-        : "all";
-    handleFilterChange("label", formattedString);
-  };
-
-  const handleSearchClick = async (e, newSearchValue) => {
-    e.preventDefault();
-    setSearchValue(newSearchValue);
-    fetchDataAndUpdateUrl();
-  };
-
-  const handleCheckboxChange = (issueId) => {
-    console.log(`Checkbox for issue ${issueId} changed.`);
-  };
-
-  const handleClearAll = () => {
-    setSelectedAuthor("all");
-    setSelectedLabel("all");
-    setSearchValue("");
-    setStateOpenOrClosed("open");
-    handleLabelChange([]);
-
-    const url = new URL(window.location.href);
-    const searchParams = new URLSearchParams();
-    searchParams.delete("q");
-    url.search = searchParams.toString();
-    window.history.pushState({}, "", url);
-
-    fetchInitialData();
-  };
+    setStateOpenOrClosed,
+    handleFilterChange,
+    handleSearchClick,
+    handleCheckboxChange,
+    handleClearAll,
+  } = useContext(IssueContext);
 
   const isDefaultState = () => {
     return (
@@ -254,8 +59,14 @@ const IssuePage = () => {
           setStateOpenOrClosed={setStateOpenOrClosed}
           authors={authors}
           labels={labels}
-          handleAuthorChange={handleAuthorChange}
-          handleLabelChange={handleLabelChange}
+          handleAuthorChange={(author) => handleFilterChange("author", author)}
+          handleLabelChange={(labels) => {
+            const formattedString =
+              labels.length > 0
+                ? labels.map((label) => `label:"${label}"`).join(" ")
+                : "all";
+            handleFilterChange("label", formattedString);
+          }}
         />
         <IssuePageList
           issuesToDisplay={apiResult}
